Rework the exports.
I don't like that the HandleScope is public.
It contains some stuff that a user should not
be able to see. Essentially it should only be
a handle with a lifetime inside the API, but
we can't just wrap it, since that would mean,
that we would need to wrap every other V8
struct that interacts with it.